Initial Commit
An initial version of the DRB ETL pipeline and API. The `dcdw` directory contains the files for a docker image which can be run in different modes to provide the different components of an application that provides all processing of incoming records. This application is designed to be run on a Kubernetes cluster.