SFR-228 Add NYPL Process and improve mapping
This adds a NYPL process that can ingest records on a daily/custom basis from the bib replica database. It also extends the data model for the DCDW to handle physicalLocation fields, which are tied generally to items, and in this case to specific links for requestable items.

These records should be tested to validate that they are being processed properly and have some initial unit testing/commenting before a PR is opened.