NOREF Handle unexpected OCLC Classify Response
This puts in a extra condition for reading the response from OCLC in case it is not one of their documented response codes. It also fixes an `if/elif` conditional error that might have been the cause of a bug in the ingest process (though it should have been fine).