NOREF Improve parsing of authors/contributors
OCLC Classify responses tend to include contributors as authors with defined roles in brackets. This attempts to identify these entries and sort them into the contributor field by looking for the presence of a bracked set of roles at the end of the author name. If found it is stripped from the name and the classify record updated accordingly.