NOREF Refactor repo to remove nested nature of Docker image
This originally had the files for the docker image in a subdirectory, but this is difficult to write tests for and seemed like it could potentially cause some confusion down the road. This unnests that directory and makes the root of the repo the root of the dockerfile.

This includes some unecessary files in the docker image but they are small and it may make debugging easier as it includes things like the CHANGELOG