SFR-951 Make S3 process more generic
This updates the functionality and some of the variable names in what was the ePubProcessor. This makes it more generic to handle any files that need to be stored and made accessible through S3. Right now the other category is covers but this might expand in the future.

All files for this project that are being stored should be publically available, which makes the storage parameters fairly simply to manage.