SFR-228 Update NYPL Ingester to check copyright
Previously the NYPL ingest process attempted to all research bib records into the DRB data warehouse. This resulted in the importing of many records that were at best tangentially related to the project as most bib records concern works published within the last 95 years.

This should speed up the ingest process as it allows us to identify the most relevant records. It also attempts to use the CCE API to identify works published between 1925 and 1964 that are in the public domain.