NOREF Fix default agent roles in ElasticSearch index
There is a shared method `addAgent` in the ElasticSearch manager that structures agent records for ES. In doing so it sets a default role for the agent if one is not provided. However this was set to be author, even though for edition records it should be publisher. This update makes that default role configurable, and therefore the method much more extensible. The default still falls back to author, since that is the most common value.