NOREF Extend RabbitMQ heartbeat
This resolves some import issues where processes were encountering issues where RabbitMQ was closing connections after an extended period of idleness. By setting a a `heartbeat` parameter on the connection we can ensure that the connection listens for a longer period of time.

In conjunction this also removes a timeout on s3File processing time to handle larger files.