SFR-956 Add PDF Manifest generator
To support the display of multi-file PDF books this generates a JSON-LD manifest that (in the future) will be read by the webreader to display the files as a seamless whole.

This constructs the manifest from extracted metadata and provides a representation that can be stored in s3 for access by other parts of the application.