SFR-956 Add Project MUSE Process
This adds a process for ingesting Project MUSE records into the DRB database. This fetches MARC records for open access books available on Project MUSE. As these books are largely available as chapter-segmented PDF files, a manifest is constructed for each book that is used to provide access directly to the book content.

This process also adds a new mapping model for `MARC` records. This mapping assumes that the source records have already been parsed by `pymarc` or a similar library, the responsibility for which lies with individual processes. This is because MARC records are generally parsed as a batch, with some difficulty lying in the separation of individual records (hence the use of an external dependency).

To construct these PDF manifests the structure is inferred from the structure of HTML pages on Project MUSE that represent these books. They are parsed using `BeautifulSoup4` and reflect the section/chapter structure as present on those pages.