SFR-952 Add DOAB Ingest Process
This adds the Directory of Open Access Books as a source of records for the DRB database. It consumes an OAI-PMH feed of MARCXML records provided by DOAB on daily/complete/custom basis.

As DOAB is comprised of records from a diverse set of publishers, this update includes specific parsers for several of the more prominent/prolific publishers to provide better access to books to users. This is done by identifying URIs from DOAB as belonging to an individual publisher and processing it accordingly, frequently providing direct access to ePub or PDF files that would not be available otherwise. Because this frequently means making `HEAD` queries against these URLs the ingest process here is somewhat slower than other ingest sources.

This also include a new arg for the main process `--singleRecord` which accepts a record identifier for a single record from an ingest source and ingests only that record. This currently is only implementd for DOAB (where it provides a useful tool for ensuring that specific books are present for testing) but is widely applicable and could certainly be implemented for the Project Gutenberg and NYPL ingest processes if desired.

Several other minor changes also go along with this and include some extensions to the PDF manifest generation methods as well as adding a new custom error type from the mapping system. This allows us to throw and catch exceptions when DCDW records are unparseable. Previously these would have either thrown an undertermined error or simply ingest data that would pollute the database.

As usual the test suite has been extended to maintain 95%+ coverage of the repo.