SFR-984 Update NYPL API token refresh method
The recommended method of obtaining `refresh_token`s from the OAUTH server does not appear to work with the NYPL server (the tokens are not returned).

With no clear way to obtain these tokens a workaround is put in place here to:

1) Catch expired token errors
2) Fetch a new token from the token endpoint
3) Use that token to re-create the API client, retry the current query and store the new client for future calls

This also includes a performance improvement that blocks the manager from repeatedly creating new clients when not necessary.