SFR-986 Update Gutenberg process to handle new media type
The new reader will not use the old JSON flags provided with each URI but instead use the media types attached as they are both easier to understand and a standard field. To support this we have settled on the `application/epub+xml` media type for the `container.xml` files that provide access to the structure of ePubs. This simply updates the Project Gutenberg process to match this new standard.