NOREF Standardize timeout and errors for DOAB
This implements a change to both set the timeout for all requests made during the DOAB process to 15 and ensuring that timeout errors are handled in all parsers. This should make the DOAB process more stable and also hopefully enable it to retrieve more files with a longer timeout value.