NOREF Allow s3 paths to be uppercase
A miscommunication in how the new webpub reader will work led to a decision to lowercase the paths of all files being stored in s3. While this is a good decision in general it breaks the internal path structure of the ePub files (which is relevant since we now serve these by individual components).

These updates allow the `META-INF` paths to be properly persisted in s3 and the database as well as ensuring that other subsidary paths are not modified in s3.