SFR-953 Add Cover fetching process
This adds a process for fetching cover images for edition records, storing them in s3 and adding this information to the edition record. Covers are sourced from 4 different projects: HathiTrust, OpenLibrary, Google Books and Baker & Taylor's ContentCafe. Because some of these sources need to have authenticated requests made the entire lifecycle for fetching covers is made within this process.

This also includes a method for resizing the cover images down to max size of 300x400 to both standardize these files and reduce the total storage size on s3. This also should make the process of serving them to users faster.

This is different from the cover fetching process in the earlier iteration of the ingest pipeline. By only fetching one cover per edition we can cut down on the processing time for this step and generally avoid duplicative work. It also means that this process does not have to interact with Redis.