SFR-960 Update Production Port
This moves the port at which the Flask application to `80` from the development port of `5000`. This works better with the ECS configuration and is likely the expected value.

This also adds an `EXPOSE` line to the Dockerfile, which is also necessary. When running the API localy it can still be access by passing `-p 5000:5000` to access the API.