NOREF Fix ElasticSearch manager when using AWS ES Service
This resolves a bug where the default python `urllib3` library is unable to parse AWS VPC URLs due to their length/unusual construction. By passing these parameters as a constructed string those checks can be bypassed and allow this manager to interact with those resources. (Which is necessary for the functioning of this cluster)