NOREF Update OCLC Catalog link handling
The OCLC Catalog mapping included a few bugs and had a few places where improvements could be made.

First, it was looking in the wrong subfield for the external resource URI (it needs to be set to subfield `u`). Second, it was not checking the first indicator for a value of `4` which denotes a reference to a resource.

The improvements are in the checking of validity of the links. This implements a new method to only allow Gutenberg, HathiTrust and InternetArchive links to be imported. Further this checks the IA links to be sure that they point to items that do not have restricted access. This method should be extended to further parse the Gutenberg and HathiTrust links as well, though that is less pressing.