SFR-994 Update RabbitMQ for remote hosts
This updates the RabbitMQ client implementation with work with a variety of more advanced settings. Originally as conceived the client simply published and read messages from the default exchange without a routing key or other common settings (due to the fact that this was all that was necessary for local development).

In deployment this will make use of a shared rabbitmq cluster and needs to be able to acces a virtual host and publish/consume messages from specified exchanges. It also needs to provide credentials to this cluster. Thesese settings were added and the necessary processes extended to use them. The client can still interact with simple RabbitMQ implementations in local or other testing environments.