NOREF Improve Env var handling
Some processes load environment variables for managers that are never initialized. This updates those managers to gracefully handle missing environment variables. They will be initialized to null, which means if they are invoked they will raise errors, which is actually an intended side effect because they should not be invoked.