SFR-1006 Update HathiTrust process
This streamlines the HathiTrust import process somewhat for memory usage. It eliminates the writing of the Hathi tsv file to a `/tmp` directory and streams it directly to the `csv` module. This shoud reduce total memory usage and coupled with other updates keep this process under the total memory threshold.