SFR-1006 Update NYPL DB query
This makes a small adjustment to the db query against the NYPL bib replica. It adds an `execution_options(stream_results=True)` to prevent the full results of the query from being loaded into memory. This prevents the memory load from the process from climbing out of control and impacting other processes.