SFR-1006 Update MUSE CSV process
The method for opening the MUSE CSV file containing record update dates was causing very high memory usage, seemingly related to opening it as a file-like `StringIO` object. A better way which provides lazy-loading of the file is to use the `requests` `iter_lines(decode_unicode=True)` method, which iterates the file without exceeding the memory thresholds.