SFR-1007 Update NYPL records
This updates NYPL bib records both on ingest and in the API response.

On ingest it updates the `media_type` values to `text/html` for catalog pages and to `application/x.html+edd` for EDD options. The second option here is temporary pending discussion and may be udpated later. This also fixes an error assigning some URLs to `nypl.com`

In the API response it sets a `location` value for item records for context with NYPL items and adds `item_id` and `edition_id` to the response for context as well