NOREF Fix Single Record Clustering
Due to a information flow misjudgement the clustering stage was skipping clustering single record works. These are valid records that have no FRBRized data because they lack adequate identifiers or simply because they are rare. This should not disqualify them from being indexed and this update ensures that they are.