SFR-1026 Add Additional Identifier to OCLC Records
This update adds the OCLC OWI number present in the Classify service to the child catalog records that are retrieved as part of that process. This ensures that there is a direct and easy connection between the OCLC Classify and OCLC Catalog records. This should both increase accuracy in matching these records and make the matching process faster by providing direct connections (rather than relying on chains of OCLC and ISBN numbers)