SFR-1026 Update record gathering process
This batches the record gathering methhod to prevent making very large and expensive queries against the `identifiers` column in the `records` table. This uses a GIN index that provides fast look ups but slows down dramatically when given large numbers of identifiers to look up.

This solution does not ultimately solve this issue but it handles it in a more graceful way to prevent blocking from occuring and ensuring that all queries will complete.