SFR-1026 Improve Grouping process
This greatly improves how records are gathered and clustered by updating the method of finding the ideal number of clusters. Instead of traversing the collection of records linerally to find this value, the set is repeatedly bisected to more quickly identify the best value. In essence this is a very rudimentary binary search to locate the best value. Coupled with previous improvements this should allow the clustering process to handle records of arbitrary size. While this can still be slow for some work records, it should be able to process them and provide a means of further improving performance down the road.