SFR-1019 Add Ingest Analytics Process
This adds a process designed to be run daily and provide an update on the number of `Work`, `Edition`, `Item` and `Record` records that were created/updated in the past 24 hours. These numbers are exported to a Smartsheet spreadsheet, where interested people can access them.

In addition the process highlights "anomalous" values where the total number of created or modified records exceeds 10% of the total number of records in the database. This is intended to both alert people to possible issues and let people who are collecting stastics that a custom job may have been run.

This is intended to solve the needs of people who are frequently asked for these numbers by outside parties.