HOTFIX Handle identifiers with commas
In rare instances an identifier may contain a comma (either intentionally or unintentionally), this can cause issues in the clustering process as the constructed query will interpret an identifier that contains a comma as two seperate values to query. This is resolved by wrapping any such identifiers in quotation marks.

This also removes a source of potential commas in identifiers by splitting hathitrust identifiers on commas on ingest. Due to the format of hathitrust records in rare cases catalogers merged multiple identifiers into a single column. This ensures that we have more accurate data with which to cluster records