SFR-1034 Add Swagger Documentation
This replaces the basic status endpoint at the API root with swagger documentation. This not only provides verification that the API is operational, but a useful guide to the endpoints and their responses. Currently this simply serves the endpoints without any prefixes but that could be updated to have a `v4` prefix.