SFR-1035 Add image block to Publications
This adds the mandatory `images` block to the OPDS2 publication model. This also attempts to populate the image block with at least one cover image sourced from the database. An error when an image is not present is currently disabled for testing purposes but will be re-enabled before merging to ensure that this complies with the OPDS2 standard