SFR-1035 Add Group object and related updates
This implements the `Group` array present in OPDS2, used when both navigation and publication arrays are present. In several cases this applied to the DRB OPDS2 feed, and this update implements those changes as well. Now search and `/new` results pages include the groups object.

This should be the last new element to be implemented to support the OPDS2 standard