SFR-1040 Convert PDF Manifests to Webpub Manifests
This performs the work of converting the custom PDF Manifest class to a class representing the standard Readium webpub manifest format. The core differences between these two formats are:

- Links are represented in Webpubs while they were an object in the custom manifest
- Webpub source files are represented through the `readingOrder` and `resources` arrays, while these were combined into the `resources` array in the custom format
- There is no representation of page numbers or spans in the webpub format

In addition to these internal formats the `media_type` will change from the custom `application/pdf+json` to the standard `application/webpub+json` which should make consuming these files easier for client applications.

Internally the Project MUSE and DOAB ingest processes had to be updated to produce these manifests, which as they are largely similar should not pose an issue.