NOREF Handle Closed Connection
In the case of encountering a closed connection when getting a message from a rabbit queue it can be assumed that the record has already been fetched. In this case we skip that record and re-create the channel. If this error was encountered incorrectly the message should still be there ready to be fetched. This should resolve possible issues with running multiple consumming applications.