NOREF Update Cluster Process committing
While clustering records, the source records must be marked as processed (to avoid duplicating work). Previously this was done with a bare `update` statement. When running concurrently with other processes this can cause a deadlock in rare cases. This method was therefore refactored so that the `commit` can happen within a `try/except` block that handles the deadlock error.