SFR-1082 Don't split queries on commas within quotation marks
To properly parse complex queries they must be split on commas. However this can cause issues when a comma is in a quoted phrase. This has two unintended consequences:

- The query is split when the user almost certainly wanted that search executed as a phrase
- The splitting results in a syntax error because it results in open quotation marks that cannot be resolved by ElasticSearch

This resolves these issues by implementing a regex to split the query string that takes quotation marks into account. Through this the semantic meaning of the quotes is preserved and the query is only split on intended commas.