SFR-1075 Filter inner_hits for sorting
When sorting on nested documents in ElasticSearch the parent documents will be returned in proper order, but any `inner_hits` documents will not automatically sorted. In a related issue, any filters applied to a search query will not automatically be applied to a sort condition. This commit tackles both of these issues with the following approach.

First, filters are created, but not yet applied to the general ElasticSearch query. Then, sort clauses are created, taking into account any edition level filters created in the previous step. This means that any response object will only include editions that match the sort criteria for the query.

Second the filters are applied to the query, at which point an `inner_hits` clause is applied with any date sort values. This is limited to date sort values because these are the only sorts that are conducted at the edition level.

This approach ensures that sort and filter values are applied in a uniform way and that no conditions are dropped. To do so the `ElasticClient` class was somewhat refactored to increase access of methods to filter and sort values generated by other methods. It also improved how the general query can be accessed, reducing the overall number of method arguments that need to be passed for many methods and improving unit testing.