SFR-1098 Filter out non-matching links
This removes links that do not match the currently configured format filter (and still return all links if no filter is set). This ensures that users see the expected results when they select a filter.

This is fairly simply due to the fact that the filter already removes any `Item` records that do not contain a valid link. So this must simply remove any other links on the records that are included.