SFR-1097 Temporarily block webpubs
In anticipation of the new webreader, many files (primarily from Project MUSE and DOAB) have been stored as Webpub manifests, which can make PDF files readable through several means. However, this reader is not yet available, meaning that these files can throw errors. This blocks them from being included in API responses until the reader is ready.

When it is, a single line (and corresponding test) will need to be removed to re-enable this functionality