HOTFIX Handle edge case error
In very rare cases a `metadata.yml` file can be generated for a Gutenberg work without a cover reference. This simply handles the error raised in that case, logs the event and continues. While we would like covers for all works, we do not consider a missing cover a reason to block ingest of a record.