HOTFIX Resolve work merging method bug
The current method for updating work records makes a faulty assumption about how the sqlalchemy `merge` method works. This is simply replaced with a different approach that deletes pre-existing records that are updated and assigns the oldest relevant created date to the newly created work record. This better reflects the process and should be faster.