SFR-1113 Fix Edition Show All Filter
In the context of an edition detail the `showAll` filter was not being passed to the proper method. Asa result the corresponding function on the front end did not update the detail page. This passes the variable to the correct function and applies the logic when iterating through the component records of an edition.