SFR-1094 Swap blocklist for allowlist for author search
The currently implemented blocklist was not properly filtering roles during author searches. This resulted in unexpected results for author searches. Because it is difficult to predict what role will be used in metadata records, this block list is replaced with an allowlist that explicitly defines the roles that are considered valid for an author search.

This is more restrictive but should be consistent in the long term regarding what is matched for such searches. The list can be extended if necessary.