SFR-1123 Update search query splitting
Previously query strings were split on commas (except for those contained in quotation marks). But this can still have unintended behavior when commas are part of a query. This resolves the issue by instead only splitting queries when a query term is present (e.g. `title:` or `subject:`). If a non-prefixed query is present or is first in a query string that is properly parsed. This should not impact any queries made by any existing applications and should only marginally impact any queries as most (if not all) users will expect this behavior.