SFR-1109 Deep paging and caching
This adds a mechanism that allows deeply paginated result sets to be retrieved. It does so by using ElasticSearch's `search_after` functionality, which allows for stateless retrieval of inner pages (though it still does scroll internally). This resolves the error with trying to retrieve results beyond the 10,000-th result, but performs very poorly in its bare form.

In addition to adding this, a caching mechanism is also introduced. This saves the `search_after` value for specific search/page combinations to speed up retrieval by up to 3x. However at present this only works for specific pages, an attempt to cache more pages couldn't be made to work asynchronously, which would be necessary to avoid a performance hit for many searches. It is also of questionable utility as many searches are unique (and deep paging is most likely a very rare use case).