SFR-1156 Replace low-level ES client with high-level
Previously this was specifically instantiating an ElasticSearch client in the low-level python library. This worked but it was unclear how this managed connections.

This is replaced with the default high-level ES client behavior, which creates a default connection that can be used by different aspects of the Elasticsearch processes. This may not resolve any issues but it is simpler, clearer and should provide a better environment especially in the API process that can spawn many concurrent connections.