NOREF Fix VIAF/LCNAF parsing
In the current clustering step records with LCNAF/VIAF numbers are being omitted from the edition-level clustering step. This is because they contain no edition-level metadata. This adds an additional step to silo these records and pull out relevant work-level metadata, most importantly the VIAF and LCNAF numbers for authors.