NOREF Update Utils endpoint with Redis client
When the redis client was added to the API process it needed to be passed wherever the ElasticSearch client is instantiated. Because it does not currently interact with the `utils/languages` endpoint it was omitted from this method, which caused an error. This adds it as well as updates the relevant unit test to verify this.