NOREF Improve Stability in processing
This implements two minor but important stability improvements. First, it extends the wait period in the catalog process. Delays in the classification process are causing this process to exit early, extending the wait process should improve this, though it may need to be further extended. Second, timeout errors are handled in the cataloging process so that the saving of elastic records is retried. This should help the process run to completion in a consistent manner.