SFR-1221 Add proxy endpoint
This adds a new endpoint that proxies requests for resources. It is intended for use with the webreader when that application has difficulty loading resources. At present it has a CORS header setting to only accept requests from `.nypl.org` addresses. This is a (probably over-cautious) security setting but shouldn't prove too difficult to work with in testing scenarios