SFR-1245 Add conformsTo for PDF Webpub Manifests
In order to facilitate the use of PDF files in Webpub Manifests, web reader clients need a method of determining the contents of a manifest. The established way of doing so is to use the `conformsTo` option in the metadata block. This implements this for PDF manifests only (as ePub manifests are assumed to be the default in the current web reader)

A PDF profile has been defined so the only udpate here is to include the relevant URIfor PDF manifests as they are generated. This is configurable via an environment variable.