SFR-1068 Collections db model
This adds a new table to the postgres model to hold collections (and associated editions). This migration is necessary to support the collections feature to be shown as a POC for further development.

To imploment this work we have added the alembic module for managing migrations. This module can be invoked via the normal process via the CLI or a container. Arbitrary arguments can be passed to upgrade/downgrade the database model as needed and as a result all future database updates will be versioned.