SFR-1246 Add Create/Delete/Retrieve Collection endpoints
This adds the basic set of endpoints for creating, retrieving and destroying collection records in the DRB database. This represents a basic CRUD system (omitting updates intentionally) to support the collections POC.

The API returns an OPDS2 object representing each collection, which contains publication records representing DRB records. Each publication contains `link` objects that can provide direct access to the resource, or the edition page within DRB that is represented by the Publication.

This also adds a test suite and other supporting tools.