NOREF Update swagger to default theme
Previously swagger was explicitly setting the scheme to be used by endpoints. Since this causes an error when loading in a different scheme (as would happen by default) this is removed in favor of using the currently deployed scheme. This works because the swagger documentation and API endpoints will always be deployed in the same manner.