NOREF Update format filtering
This updates the format filtering process to account for webpub files, which do not share a `media_type` with the files that they wrap. However we can rely on the fact that different file types are considered different items in the DRB schema. As a result, if after filtering we are left with only a webpub link for an item, we can consider that the wrong format and exclude it