HOTFIX Improve Record Handling for new reader
This implements a simple fix for the proxy endpoint that handles relative redirects from source files. It resolves these using the extracted scheme and netloc of the original url, which must contain these values.

In addition this tweaks the settings for the clustering process to produce more accurate work records.