SFR-1304 Add highlighting to search responses
This adds the highlighted fields from ElasticSearch to the DRB API response for both the default DRB API format and the OPDS2 feed. This provides information to consuming applications about what strings, in what fields, were matched against.

This is returned in a new `_meta` object in both APIs, an object is designed to be extensible and contain future metadata about DRB records in the future (prior to this all metadata pertained to the works being described by the DRB records).