NOREF Fix format filtering bug
An issue when filtering with formats was due to a conflict with how webpub manifests must be handled. This removes that format from basic filtering and adds it back to criteria when necessary. This simplifies the process and ensures that the filters work conceptually as they would be expected to.