NOREF Sort links by media_type
Sorting links in the API by `media_type` provides an extra layer of assurance that deplyoments while we are transitioning to a new version of the web reader will not cause errors. It ensures that the "safest" links are displayed first, while those that provide more advanced features are displayed last. This can be reversed to improve performance after web reader deployment