WIP Language Detection and Indexing
This is a work-in-progess commit of the language detection and indexing implementation. It uses a pipeline that sits in front of the ES saving process. It requires an ES7 cluster to use this functionality and does not have a fully realized ES mapping to sit behind this process. Also further work will need to be done to set up the language detection pipeline to operate on all desired fields.