SFR-1387 Improve proxy redirects
In rare cases (from a subset of publishers) `302` redirects may exhibit odd behavior that prevents them from being handled in a standard way. This inserts a fallback that relies on the `requests` module parsing of the redirect. While generally this can have issues, these edge case redirects are accurately handled by the module (as it more closely represents a browser environment).