SFR-1440 Remove Polyglot
This removes Polyglot, which was previously used for language detection at the OCLC Classify phase. This was a good tool but has not been actively maintained in some time and has experienced enough dependency drift that it has broken our docker builds. With little confidence in this changing this should be replaced.

The `fasttext` is more widely adopted, is under active maintenance, and is roughly as fast. It provides accurate detection when the `threshold` value is provided. In the current code this is set to `0.5` as this is a relatively low bar and we want to ensure that most strings are assigned a language.

One test was added for non-roman script testing to increase confidence in the tool. This should be a straight swap of the dependency so we are not expecting any carry on effects from this change.