SFR-1456 Add ES Retry Functionality
Due to an issue in the ElasticSearch 8.x python client, we cannot control the timeout values for bulk operations in ES. This is an occasional issue that crops up with large batches of records.

This fix adds a split and retry functionality for batches, catching `ConnectionTimeout` errors, splitting the offending batch in half and retrying each recursively.