SFR-1460 Update script running approach
To improve the ease which we can run scripts this adds a `scriptRunner` file and makes a few other improvements. To run a one-off script the process will be:

1. Write the script in a module in the `scripts/` directory. This should have one method that can be invoked to run the script
2. Export the script in the `scripts/__init__.py` module with an accessible name (e.g. `from .testScriptFile import scriptFunc as testScript`
3. Run the script with `python scriptRunner.py testScript`

This also adds a minor update to the edition extraction script to handle the case where the language identified in a work record doesn't have a translation available in the file.