SFR-1498 Add ES Index updating script
This script finds all work/edition records that have a resolvable link pointing to an NYPL resource. If that resource is a link to a catalog record (not to an EDD option), the `mediaType` of that link is updated to `application/html+catalog` which prevents these links from being picked up as readable.