SFR-1513 Update Ingest report query strategy
The ingest query was using a `casting` approach to convert created/modified dates from `datetime` to `date`. This had to be done on individual rows and slowed the query to the point that would cause it to timeout.

This updates the approach to remove the casting and updates the queries to run a more standard date range query.